cambio tipo datos double

+++ microservicio/dominio/src/main/java/com/ceiba/pedido/modelo/dto/DtoPedido.java
@@ -15,8 +15,8 @@ public class DtoPedido {
     private Long comprador;
     private String ciudad;
     private String destino;
-    private int flete;
-    private int precio;
-    private int descuento;
-    private int total;
+    private double flete;
+    private double precio;
+    private double descuento;
+    private double total;
 }
